Show working tickets
Added function to show all working tickets with summary data.

Also added function to get info about a user by user-id, since this is
the only information about the requester provided in the individual
ticket hash.

Currently the only use for the user-id lookup function is to get the
name of the user that requested a ticket, but you should be able to use
it to look up any information about a user.